requirements.txt
matplotlib==3.5.0
Pillow==9.0.1
scikit_learn==1.0.2
streamlit==1.5.0
